src/requirements.txt
flet
minecraft-launcher-lib
install-jdk
wget
numpy
attrs
pillow